Rename array utility files to lowercase
